+++ src/systems/experimental/mod.rs
@@ -1,2 +1,2 @@
-pub mod flocking;
+pub mod boids;
 pub mod perception;
